Add ProgStudentApp project and extend WorkerAppV0 functionality
A new project named "ProgStudentApp" has been added to the Visual Studio solution `WorkerAppV0.sln`. This includes the addition of the project configuration settings for different build configurations (Debug|x64, Debug|x86, Release|x64, Release|x86).

In `CWorker.cpp`, a closing brace `}` was added to the `IncreaseRate` method.

In `WorkerAppV0.cpp`, the `main` function was extended to include functionality for checking if the user wants to increase the hourly rate for a worker. This includes prompting the user for a worker number, displaying worker details, and asking if they want to offer a percentage increase in pay.

A new file `CProgStudent.cpp` was created and includes the `CProgStudent.h` header file.

A new header file `CProgStudent.h` was created, defining a `CProgStudent` class with private member variables for name, work number, hourly rate, and hours worked.

A new source file `ProgStudentApp.cpp` was created, which includes the `main` function for the `ProgStudentApp` project. Currently, the `main` function is empty.

A new project file `ProgStudentApp.vcxproj` was created, defining the project structure, configurations, and settings for the `ProgStudentApp` project.

A new filters file `ProgStudentApp.vcxproj.filters` was created, organizing the source and header files into appropriate filters for the `ProgStudentApp` project.